backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
langchain==0.3.13
langchain_community==0.3.13
langchain_core==0.3.28
langchain_openai==0.2.14
langchain_pinecone==0.2.0
langchain_text_splitters==0.3.4
python-dotenv==1.0.1
pinecone-client==5.0.1
pinecone==5.3.1
pypdf==5.0.1
notebook==7.2.2
black==24.8.0
requests==2.31.0
pywhatkit==5.4
qrcode==7.4.2